+++ cmd/juju/romulus/showwallet/show_wallet.go
@@ -12,8 +12,8 @@ import (
 	"github.com/juju/errors"
 	"github.com/juju/gnuflag"
 	"github.com/juju/loggo"
-	api "github.com/juju/romulus/api/wallet"
-	wireformat "github.com/juju/romulus/wireformat/wallet"
+	api "github.com/juju/romulus/api/budget"
+	wireformat "github.com/juju/romulus/wireformat/budget"
 	"gopkg.in/juju/names.v2"
 	"gopkg.in/macaroon-bakery.v1/httpbakery"
